Add files via upload
Many royal inscriptions are found on multiple objects and represent multiple documents, albeit documents that contain exactly the same inscription. input_multiple contains the same files as input, but duplicated where necessary to reflect this reality. For example, if an inscription was found on 10 individual objects, then there are now 10 copies of that inscription in input_multiple instead of just one.